refactor(Result): domain.game 패키지로 이동
