+++ src/main/java/domain/game/BlackjackGameManager.java
@@ -1,6 +1,5 @@
 package domain.game;
 
-import constant.Result;
 import domain.card.Card;
 import domain.card.CardMachine;
 import domain.participant.Dealer;
