+++ src/main/java/domain/game/Result.java
@@ -1,4 +1,4 @@
-package constant;
+package domain.game;
 
 public enum Result {
     WIN("승"),
